pyproject.toml
# Add basic project information and dependencies
[project]
name = "fleet-mix-app"  # Name must not contain spaces.
dynamic = ["version"]
dependencies = [
    "flask",
    "pandas",
    "plotly",
    "pytest",
    "pytest-cov",
    "pytest-playwright",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

